# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/lectures": {
            "get": {
                "tags": [
                    "lectures"
                ],
                "description": "Returns lectures",
                "operationId": "findLectures",
                "parameters": [
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "sorts items base of field provided",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "id",
                                "-id",
                                "created_at",
                                "-created_at",
                                "course_id",
                                "-course_id",
                                "-location",
                                "location",
                                "day_id",
                                "-day_id",
                                "period_id",
                                "-period_id"
                            ]
                        },
                        "examples": {
                            "-id": {
                                "summary": "sorts in an descending manner by append a '-' ",
                                "value": "-id"
                            },
                            "id": {
                                "summary": "sorts in an ascending manner",
                                "value": "id"
                            }
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "filters base on the parameter passed e.g filter[field]=value",
                        "required": false,
                        "style": "deepObject",
                        "schema": {
                            "properties": {
                                "course_id": {
                                    "type": "array",
                                    "items": {
                                        "type": "number",
                                        "example": "1"
                                    }
                                },
                                "period_id": {
                                    "type": "array",
                                    "items": {
                                        "type": "number",
                                        "example": "1"
                                    }
                                },
                                "day_id": {
                                    "type": "array",
                                    "items": {
                                        "type": "number",
                                        "example": "1"
                                    }
                                },
                                "location": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "nlt"
                                    }
                                },
                                "lecturer": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "lecturer a"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    },
                    {
                        "name": "include",
                        "in": "query",
                        "description": "includes the details of the field included e.g include='course'",
                        "required": false,
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "enum": [
                                    "day",
                                    "period",
                                    "course"
                                ]
                            }
                        }
                    },
                    {
                        "$ref": "#components/parameters/range"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success respones",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Course"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    }
                }
            }
        },
        "/api/lecture/{id}": {
            "get": {
                "tags": [
                    "lectures"
                ],
                "description": "Returns a material based on a single ID",
                "operationId": "findLectureById",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Lecture"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                }
            }
        },
        "/api/materials": {
            "get": {
                "tags": [
                    "materials"
                ],
                "description": "Returns materials",
                "operationId": "findMaterials",
                "parameters": [
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "sorts items base of field provided",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "id",
                                "-id",
                                "created_at",
                                "-created_at",
                                "course_id",
                                "-course_id",
                                "-title",
                                "title"
                            ]
                        },
                        "examples": {
                            "-id": {
                                "summary": "sorts in an descending manner by append a '-' ",
                                "value": "-id"
                            },
                            "id": {
                                "summary": "sorts in an ascending manner",
                                "value": "id"
                            }
                        }
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "description": "filters base on the parameter passed e.g filter[field]=value",
                        "required": false,
                        "style": "deepObject",
                        "schema": {
                            "properties": {
                                "course_id": {
                                    "type": "array",
                                    "items": {
                                        "type": "number",
                                        "example": "1"
                                    }
                                },
                                "title": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "mat101"
                                    }
                                },
                                "description": {
                                    "type": "array",
                                    "items": {
                                        "type": "string",
                                        "example": "pdf"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    },
                    {
                        "$ref": "#components/parameters/range"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Material"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/400"
                    }
                }
            }
        },
        "/api/materials/{id}": {
            "get": {
                "tags": [
                    "materials"
                ],
                "description": "Returns a material based on a single ID",
                "operationId": "findMaterialById",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "single material response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Material"
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/404"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Course": {
                "required": [
                    "code",
                    "level_id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "description": "the course title or course name",
                        "type": "string"
                    },
                    "code": {
                        "description": "the course code",
                        "type": "string"
                    },
                    "level_id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Course"
                }
            },
...
            "Lecture": {
                "required": [
                    "course_id",
                    "day_id",
                    "period_id",
                    "location",
                    "lecturer"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "course_id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "day_id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "period_id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "location": {
                        "type": "string"
                    },
                    "lecturer": {
                        "type": "string"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Lecture"
                }
            },
...
            "Material": {
                "required": [
                    "course_id",
                    "title",
                    "description",
                    "url"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "course_id": {
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/created_at"
                    },
                    "updated_at": {
                        "$ref": "#/components/schemas/BaseModel/properties/updated_at"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Material"
                }
            },
...
        },
        "responses": {
            "404": {
                "description": "not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": ""
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "400": {
                "description": "Error: Bad Request",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Requested filter(s) `description, are not allowed. Allowed filter(s) are `created_at` "
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "parameters": {
...
            "id": {
                "name": "id",
                "in": "path",
                "description": "ID of model to fetch",
                "required": true,
                "schema": {
                    "type": "integer",
                    "format": "int64"
                }
            }
        }
    },
    "security": [
        []
    ]
}